pyproject.toml
[project]
name = "vibeweb-socketio"
version = "0.1.0"
description = "Production-ready SocketIO server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-socketio>=5.11.0",
    "uvicorn[standard]>=0.27.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
]
dashboard = [
    "reflex>=0.5.0",
    "python-socketio[client]>=5.11.0",
]

[project.scripts]
server = "app.main:run_server"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
package = true
